Add test code for ox-asciidoc
ox-asciidoc is still small piece of code but it already passed my
ability to check all of the combination of transcode pattern with
local test files. ERT is much easier to cover behaviors of the
ox-asciidoc.el compare to those test files.

This is a first step toward TDD with ERT.