Convert example blocks and fixed-width to the AsciiDoc's Literal Blocks
Asciidoc differentiates Listing Blocks ("----") and Literal
Blocks ("...."), however the current ox-asciidoc converts all three
elements, src block, example block, and fixed width, to the Listing
Block.

This patch changes the callback functions for the example block and the
fixed width to output the Literal Block delimiter lines ("...."), but
keep src blocks to the Listing Blocks ("----").

ox-html converts all three elements to pre tags, but classifies the
example blocks and the fixed-widths as "example" and the src blocks to
"src".

This closes #6.