Add List continuation and discontinuation support
A list item can be arbitrary object, including multiple lines of
paragraph.  In Org, you can put any number of item under one list item.
However, In Asciidoc, you can only put one item.

In Asciidoc, when you want to put multiple paragraphs as a list item,
you connects them with `+', as list continuation.

In contrast, Org's list ends with either a less indented line or two
blank lines.  In this case, you also need to end the list in Asciidoc.
This can be possible with `// ^', a comment line with a dummy comment
string `^'.

This closes #7.