Add support for Asciidoctor diagram
Asciidoctor diagram is one of the official extensions of Asciidoctor.
It uses a special block for each diagram to post-process just like Org
Babel does.  In order to support it, we have to export a source code
block as a special block with "[" diagram-type "]"; eg) [ditaa] for
ditaa.

This commit is an attempt to support it with only one test case.

This closes #14.

https://docs.asciidoctor.org/diagram-extension/latest/

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>